apps/cli-client/pyproject.toml
[tool.poetry]
name = "photos_drive"
version = "8.0.6"
description = ""
authors = ["Emilio Kartono <e.kartonoe@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
exifread = "^3.0.0"
typer = "^0.15.1"
backoff = "^2.2.1"
types-requests = "^2.32.0.20241016"
google-api-python-client = "^2.156.0"
google-auth-oauthlib = "^1.2.1"
python-magic = "^0.4.27"
tqdm = "^4.67.1"
event-bus = "^1.0.2"
dacite = "^1.8.1"
types-exifread = "^3.0.0.20240806"
pymongo = "<4.11"
prettytable = "^3.12.0"
termcolor = "^2.5.0"
xxhash = "^3.5.0"
pyexiftool = "^0.5.6"
pillow = "^11.2.1"
pillow-heif = "^0.22.0"
opencv-python-headless = "^4.11.0.86"
h3 = "^4.3.0"
langchain = "^0.3.26"
faiss-cpu = "^1.11.0.post1"
langchain-google-vertexai = "^2.0.27"
langchain-community = "^0.3.27"
langchain-experimental = "^0.3.4"
torch = "=2.2.2"
numpy = "<2"
open-clip-torch = "^2.32.0"
langchain-google-genai = "^2.1.8"
transformers = "^4.53.2"
pydantic = "^2.11.7"
langgraph = "^0.5.4"
langmem = "^0.0.28"
rich = "^14.0.0"

[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
mypy = "^1.13.0"
pytest = "^8.3.4"
flake8 = "^7.1.1"
flake8-bugbear = "^24.12.12"
coverage = "^7.6.9"
pytest-mock = "^3.14.0"
requests-mock = "^1.12.1"
freezegun = "^1.5.1"
mongomock = "^4.3.0"
pyfakefs = "^5.7.4"
unittest-parametrize = "^1.6.0"
isort = "^6.0.1"

[[tool.poetry.source]]
name = "torch-cpu"
url = "https://download.pytorch.org/whl/cpu"
priority = "explicit"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
photos_drive_cli = "photos_drive.cli.app:main"

[[tool.mypy.overrides]]
module = "google_auth_oauthlib.*"

[tool.mypy]
check_untyped_defs = true
ignore_missing_imports = true
mypy_path = "typings"

[tool.coverage.run]
omit = [".*", "*/site-packages/*", "config.py", "config-3.py"]

[tool.coverage.report]
fail_under = 50

[tool.black]
line-length = 88
target-version = ['py37']
skip-string-normalization = true

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_sort_within_sections = true
